Rename service.ts to CrSql.ts and update exports
